--- src/client/config.rs
@@ -1,6 +1,0 @@
-pub struct Config {
-    pub app_id: String,
-    pub cluster: String,
-    pub servers: Vec<String>,
-    pub secret: String,
-}
